Stuff
Some stuff I've been working on. Note that please check before opening the death code files, they are really dangerous.
The a test.bat program downloads the script for creating the memz virus from a website, converte it into a batch file, save it to the downloads folder on your machine and with a seperate file attached that you can click to start the program. Note that the program only run successfully with windows security's real time protetion and other virus protections turned off, since it can harm your computer and damage the boot drive.